+++ assets/javascripts/kitten/karl/cards/donation-card.js
@@ -1,22 +1,13 @@
-import React from 'react'
-import Radium, { StyleRoot } from 'radium'
+import React, { Fragment } from 'react'
 import { DonationCard } from 'kitten/components/cards/donation-card'
 import { Marger } from 'kitten/components/layout/marger'
-import {
-  Grid as GridBase,
-  GridCol as GridColBase,
-} from 'kitten/components/grid/grid'
-import { TextInputWithUnit as TextInputWithUnitBase } from 'kitten/components/form/text-input-with-unit'
-import { Text as TextBase } from 'kitten/components/typography/text'
+import { Grid, GridCol } from 'kitten/components/grid/grid'
+import { TextInputWithUnit } from 'kitten/components/form/text-input-with-unit'
+import { Text } from 'kitten/components/typography/text'
 import { Container } from 'kitten/components/grid/container'
 import { Label } from 'kitten/components/form/label'
 import { KarlExampleTitle } from 'kitten/karl/examples/title'
 import { mediaQueries } from 'kitten/hoc/media-queries'
-
-const Grid = Radium(GridBase)
-const GridCol = Radium(GridColBase)
-const Text = Radium(TextBase)
-const TextInputWithUnit = Radium(TextInputWithUnitBase)
 
 const KarlDonationInputBase = props => {
   const {
@@ -66,7 +57,7 @@ const KarlDonationInputBase = props => {
 }
 
 export const KarlDonationCard = () => (
-  <StyleRoot>
+  <Fragment>
     <KarlExampleTitle>Donation card</KarlExampleTitle>
     <Container>
       <Marger top="2" bottom="2">
@@ -125,9 +116,9 @@ export const KarlDonationCard = () => (
         />
       </Marger>
     </Container>
-  </StyleRoot>
+  </Fragment>
 )
 
-export const KarlDonationInput = mediaQueries(Radium(KarlDonationInputBase), {
+export const KarlDonationInput = mediaQueries(KarlDonationInputBase, {
   viewportIsSOrLess: true,
 })