+++ spec/dummy/client/entries/app-kitten.js
@@ -39,6 +39,7 @@ import {
 } from 'kitten/karl/buttons/button-with-icon-words'
 import { KarlPaymentButtons } from 'kitten/karl/buttons/payment-button'
 import { KarlButtonIconFilter } from 'kitten/karl/buttons/button-filter'
+import { KarlGarbageButton } from 'kitten/karl/buttons/garbage-button'
 
 // Cards
 import { ArticleCard } from 'kitten/components/cards/article-card'
@@ -51,6 +52,7 @@ import { ProjectSimilarCard }
   from 'kitten/components/cards/project-similar-card'
 import { SimpleCard } from 'kitten/components/cards/simple-card'
 import { KarlCard } from 'kitten/karl/cards/card'
+import { KarlDonationCard } from 'kitten/karl/cards/donation-card'
 import { KarlRewardCard } from 'kitten/karl/cards/reward-card'
 import {
   KarlSimilarProjectsCard,
@@ -350,6 +352,7 @@ ReactOnRails.register({
   KarlLinkedinButtonIconWords,
   KarlTwitterButtonIconWords,
   KarlButtonIconFilter,
+  KarlGarbageButton,
 
   // Colors
   KarlColors,
@@ -363,6 +366,7 @@ ReactOnRails.register({
   ProjectSimilarCard,
   SimpleCard,
   KarlCard,
+  KarlDonationCard,
   KarlRewardCard,
   KarlSimilarProjectsCard,
   KarlSimilarProjectsWithInfosCard,