+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add social button icon component `YoutubeButtonIcon` and add `YoutubeIcon` component.
+- Fix: Update sizes on all social `buttonIcon`.
 
 ## [20.12.3] - 2018-05-22